+++ re_contact_book_system/backend/migrations/versions/7e216e07dc5e_initial_tables.py
@@ -1,5 +1,6 @@
 from alembic import op
 import sqlalchemy as sa
+from sqlalchemy.sql import text
 
 
 revision = "xxxx_initial_tables"